enable background script to run the fetch requests
