+++ browser-extension/manifest.json
@@ -15,37 +15,6 @@
     "type": "module"
   },
 
-  "content_scripts": [
-    {
-      "matches": [
-        "https://training.vri-research.com/index.php/admin/edispnew/sa/index/surveyid/*",
-        "https://training.vri-research.com/index.php/admin/responses/sa/*/surveyid/*",
-
-        "https://opinion-insight.com/index.php/admin/edispnew/sa/index/surveyid/*",
-        "https://opinion-insight.com/index.php/admin/responses/sa/*/surveyid/*",
-
-        "https://research-opinions.com/index.php/admin/edispnew/sa/index/surveyid/*",
-        "https://research-opinions.com/index.php/admin/responses/sa/*/surveyid/*",
-
-        "https://opinions-survey.com/index.php/admin/edispnew/sa/index/surveyid/*",
-        "https://opinions-survey.com/index.php/admin/responses/sa/*/surveyid/*",
-
-        "https://insight-polls.com/index.php/admin/edispnew/sa/index/surveyid/*",
-        "https://insight-polls.com/index.php/admin/responses/sa/*/surveyid/*",
-
-        "https://viewpointpoll.com/index.php/admin/edispnew/sa/index/surveyid/*",
-        "https://viewpointpoll.com/index.php/admin/responses/sa/*/surveyid/*",
-
-        "https://p.vri-research.com/index.php/admin/edispnew/sa/index/surveyid/*",
-        "https://p.vri-research.com/index.php/admin/responses/sa/*/surveyid/*",
-
-        "https://m.vri-research.com/index.php/admin/edispnew/sa/index/surveyid/*",
-        "https://m.vri-research.com/index.php/admin/responses/sa/*/surveyid/*"
-      ],
-      "js": ["test.js"]
-    }
-  ],
-
   "permissions": [
     "activeTab",
     "tabs",
